Set a static packageGuid
Why is this change needed?
--------------------------
Without setting this, the generator makes a new GUID on every
generation. This is annoying because it shows up in the `git diff`. This
GUID is related to the project, not the particular version of the
project[1], so it shouldn't need to change over time.

How does it address the issue?
------------------------------
Preserve the GUID from the last public release so it will continue to be
used in the future.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
[1] - https://stackoverflow.com/a/72174824

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?