Update to openapi-generator 6.6.0
Why is this change needed?
--------------------------
In general we'd like to stay up to date, but in this particular case I
plan to file a bug report with openapi-generator and I want to make sure
that the bug is not already fixed in a newer version.

How does it address the issue?
------------------------------
Update to latest released version.

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?