Fix issue with requiring document_content
Why is this change needed?
--------------------------
Before this change the document_content was required, but it should not
be required if document_url is provided. This prevented the use of
document_url.

How does it address the issue?
------------------------------
Stop the client from requiring document_content. The API will still
require it if document_url is not provided.

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?