Fix issue with passing PrinceOptions
Why is this change needed?
--------------------------
Without this the following code will not work:

```csharp
    doc = new Doc(
      name: "csharp-sync.pdf",
      test: true,
      documentUrl: "https://expired.badssl.com/",
      documentType: Doc.DocumentTypeEnum.Pdf,
      princeOptions: new PrinceOptions(
        insecure: true
      )
    );
```
and will result in the following error (assuming `new PrinceOptions` is
on line 43):
```
test/prince_options.cs(43,22): error CS1729: The type `DocRaptor.Model.PrinceOptions' does not contain a constructor that takes `24' arguments
/app/bin/DocRaptorWithDependencies.dll (Location of the symbol related to previous error)
Compilation failed: 1 error(s), 0 warnings
```

How does it address the issue?
------------------------------
Switching the argument default to `null` prevents the error. This should
be fine because the constructor checks if this argument is null and
defaults to `InputEnum.Html` if it is.

It's not clear to me why this errors in the first place, it seems like
using an enum as a default argument should be fine, but I'm not super
familiar with C#.

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?