Add a test for PrinceOptions
Why is this change needed?
--------------------------
We want to ensure that PrinceOptions continues to work as expected.

How does it address the issue?
------------------------------
Add a test that passes PrinceOptions, and verifies that the passed
options have the intended effect.

In this case we test using the prince_options[insecure]=true option
while using a url with bad SSL. If you don't pass the option, the
request will fail with an SSL error. If you do pass the option, the
request will succeed.

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?