Add a patch to fix PrinceOptions issues
Why is this change needed?
--------------------------
Adding this patch means that regenerating the client will not break the
ability to pass PrinceOptions to the Doc constructor.

How does it address the issue?
------------------------------
Add a patch to the patches directory, which will be applied to the
generated client after generation.

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?