Switch to more explicit docker detection
Why is this change needed?
--------------------------
I believe I had an issue with one of the clients where the previous
check was ambiguous inside the container. In any case this is more
explicit.

How does it address the issue?
------------------------------
Use the INSIDE_DOCKER env var from .docker_env.list to determine if
we are inside the container.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29007795/todolists/5257298600

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?