Improve generated file review
Why is this change needed?
--------------------------
These files might have useful changes that we want to review.

How does it address the issue?
------------------------------
Commit these files to the review directory so we'll be able to more
easily review changes in the future.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29007795/todolists/5257298600

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?