Remove old broken runtime environments
Why is this change needed?
--------------------------
It appears these old images use old versions of debian and `apt update`
doesn't work anymore.

How does it address the issue?
------------------------------
Remove the broken images.

It might be better in general to actually build our images and upload
them to docker hub. That way we can freeze the content of the image,
including all necessary dependencies, and not have to worry about
them changing out from under us. This'll be a task for another day,
though.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29007795/todolists/5257298600

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?