Fix issues with sending default false values
Why is this change needed?
--------------------------
When working on adding the iframes option I discovered that the client
wouldn't actually send the value for `prince_options[iframes]=false`.
This was because the default value for the property was false and the
client was not sending any values that were the same as the default.

How does it address the issue?
------------------------------
This updates to the openapi 3.0.0 spec, which allows setting a model
attribute as nullable. This allows the client to send the value `null`
as a default, indicating to the server that the server should use a
default value.

This commit just has the updated spec and regenrated client. A future
commit will add the new options.

This also removes the patch we had because we work around the issue by
not setting a default value for the `prince_options[input]` property.

This also tweaks a few other properties in the spec so values that don't
need to be sent to the server won't be.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29007795/todolists/5257298600

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?