Draft v4.0.0 release changes
Why is this change needed?
--------------------------
We'd like to release a new version to fix some bugs and add some API
options.

How does it address the issue?
------------------------------
Bump the version number and update the changelog.

I'm bumping the major version because the upgrade to openapi-generator
v6.6.0 and the new openapi spec mean some default API option behavior
has changed. I'm not sure if customers rely on the old behavior, but
I'd rather be safe than sorry.

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?